## Summary

Implement a customizable key selection strategy for Bifrost by introducing a KeySelector interface that allows users to define their own logic for selecting API keys.

## Changes

- Added a `KeySelector` function type in schemas to allow custom key selection strategies
- Added a `keySelector` field to the Bifrost struct to store the custom selector
- Extracted the default weighted random selection logic into a standalone `WeightedRandomKeySelector` function
- Added the selected key ID to the request context for tracking and debugging purposes
- Made the key selection configurable through the BifrostConfig

## Type of change

- [x] Feature
- [ ] Bug fix
- [ ] Refactor
- [ ] Documentation
- [ ] Chore/CI

## Affected areas

- [x] Core (Go)
- [ ] Transports (HTTP)
- [ ] Providers/Integrations
- [ ] Plugins
- [ ] UI (Next.js)
- [ ] Docs

## How to test

```sh
# Core/Transports
go version
go test ./...
```

You can test by implementing a custom key selector and passing it to the BifrostConfig:

```go
customSelector := func(ctx *context.Context, keys []schemas.Key, providerKey schemas.ModelProvider, model string) (schemas.Key, error) {
    // Your custom selection logic here
    return keys[0], nil
}

config := schemas.BifrostConfig{
    // Other config...
    KeySelector: customSelector,
}

bifrost, err := core.Init(context.Background(), config)
```

## Breaking changes

- [ ] Yes
- [x] No

## Related issues

Enables more flexible key selection strategies for different use cases like round-robin, least-used, or priority-based selection.

## Security considerations

The key selection strategy could potentially impact rate limiting and usage patterns across API keys, but doesn't introduce new security concerns.

## Checklist

- [x] I added/updated tests where appropriate
- [x] I verified builds succeed (Go and UI)

+++ framework/streaming/transcription.go
@@ -66,7 +66,7 @@ func (a *Accumulator) processAccumulatedTranscriptionStreamingChunks(requestID s
 	if accumulator.StartTimestamp.IsZero() || accumulator.FinalTimestamp.IsZero() {
 		data.Latency = 0
 	} else {
-		data.Latency = float64(accumulator.FinalTimestamp.Sub(accumulator.StartTimestamp).Nanoseconds()) / 1e6
+		data.Latency = accumulator.FinalTimestamp.Sub(accumulator.StartTimestamp).Nanoseconds() / 1e6
 	}
 	data.EndTimestamp = accumulator.FinalTimestamp
 	data.TranscriptionOutput = completeMessage
